style: move token counter to its own class
